refactor: 메소드 및 클래스 분리
