+++ README.md
@@ -12,8 +12,8 @@
 * [x] 빈 문자열이나 null을 입력할 경우 0을 반환해야 한다.
 * [x] 숫자 하나만 들어올 경우 그 숫자를 반환한다.
 * [x] 구분자는 기본적으로 ',', ':' 을 사용할 수 있다.
-* [ ]  “//”와 “\n” 문자 사이에 커스텀 구분자를 지정할 수 있다. (예 : “//;\n1;2;3” => 6)
-* [ ] 음수를 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “-1,2,3”)
+* [x]  “//”와 “\n” 문자 사이에 커스텀 구분자를 지정할 수 있다. (예 : “//;\n1;2;3” => 6)
+* [x] 음수를 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “-1,2,3”)
