+++ src/main/java/calculator/domain/StringCalculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import java.util.Arrays;
 import java.util.List;
@@ -6,17 +6,20 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class StringCalculator {
+
+    public static final String MESSAGE = "음수가 포함되어 있습니다.";
+
     public static int splitAndSum(String expression) {
         if (expression == null || expression.isEmpty()) {
             return 0;
         }
         List<String> numbers = splitNumbers(expression);
         if (numbers.stream().anyMatch(x -> Integer.parseInt(x) < 0)) {
-            throw new RuntimeException("음수가 포함되어 있습니다.");
+            throw new RuntimeException(MESSAGE);
         }
-            return numbers.stream()
-                    .mapToInt(Integer::parseInt)
-                    .sum();
+        return numbers.stream()
+                .mapToInt(Integer::parseInt)
+                .sum();
     }
 
     private static List<String> splitNumbers(String expression) {
