+++ src/main/java/calculator/view/ConsoleOutput.java
@@ -1,0 +1,4 @@
+package calculator.view;
+
+public class ConsoleOutput {
+}
