+++ src/main/java/calculator/view/UserInput.java
@@ -1,0 +1,10 @@
+package calculator.view;
+
+import java.util.Scanner;
+
+public class UserInput {
+    static final Scanner scanner = new Scanner(System.in);
+    public static String inputExpression() {
+        return scanner.nextLine();
+    }
+}
